// src/api-serverless/src/generated/models/DropMedia.ts
/**
 * Seize API
 * Following is the API interface description for most common operations in Seize API. Some modifying endpoints may need authentication token.
 *
 * OpenAPI spec version: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class DropMedia {
  'url': string;
  'mime_type': string;

  static readonly discriminator: string | undefined = undefined;

  static readonly attributeTypeMap: Array<{
    name: string;
    baseName: string;
    type: string;
    format: string;
  }> = [
    {
      name: 'url',
      baseName: 'url',
      type: 'string',
      format: ''
    },
    {
      name: 'mime_type',
      baseName: 'mime_type',
      type: 'string',
      format: ''
    }
  ];

  static getAttributeTypeMap() {
    return DropMedia.attributeTypeMap;
  }

  public constructor() {}
}