// src/api-serverless/src/generated/models/DropMetadata.ts
/**
 * Seize API
 * Following is the API interface description for most common operations in Seize API. Some modifying endpoints may need authentication token.
 *
 * OpenAPI spec version: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class DropMetadata {
  'data_key': string;
  'data_value': string;

  static readonly discriminator: string | undefined = undefined;

  static readonly attributeTypeMap: Array<{
    name: string;
    baseName: string;
    type: string;
    format: string;
  }> = [
    {
      name: 'data_key',
      baseName: 'data_key',
      type: 'string',
      format: ''
    },
    {
      name: 'data_value',
      baseName: 'data_value',
      type: 'string',
      format: ''
    }
  ];

  static getAttributeTypeMap() {
    return DropMetadata.attributeTypeMap;
  }

  public constructor() {}
}