// src/api-serverless/src/generated/models/DropRater.ts
/**
 * Seize API
 * Following is the API interface description for most common operations in Seize API. Some modifying endpoints may need authentication token.
 *
 * OpenAPI spec version: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { ProfileMin } from '../models/ProfileMin';
import { HttpFile } from '../http/http';

export class DropRater {
  'profile': ProfileMin;
  'rating': number;

  static readonly discriminator: string | undefined = undefined;

  static readonly attributeTypeMap: Array<{
    name: string;
    baseName: string;
    type: string;
    format: string;
  }> = [
    {
      name: 'profile',
      baseName: 'profile',
      type: 'ProfileMin',
      format: ''
    },
    {
      name: 'rating',
      baseName: 'rating',
      type: 'number',
      format: 'int64'
    }
  ];

  static getAttributeTypeMap() {
    return DropRater.attributeTypeMap;
  }

  public constructor() {}
}